requirements.txt
python-dotenv==1.0.1
fastapi==0.111.1
uvicorn==0.30.3
pymongo==4.10.1
pydantic==2.8.2
pyjwt==2.8.0

opencv-contrib-python==4.10.0.84
numpy==2.1.2
pillow==11.0.0